temporary loadout tweaks until balance is Manifested (#21)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

<!-- NOTE: This format is NOT REQUIRED for things like runtime fixes,
code fixes and optimizations. In those instances you should try to give
a description of what is being fixed or optimized but are not required
to fill out the complete changelog. -->
<!-- Adjusting things like armor or weapon values for balance, while
they may be 'fixes' in their own way, are not considered code fixes as
described above and require the use of the Pull Request format below.-->


## About The Pull Request

there will be a full balance PR hashed out with the assistance of raptor
in the coming days, however for now this is a copium patch to get the
server balance in a not-horrible state for server launch

## Why It's Good For The Game

lowers the amount of PA and nukes Insane Guns in loadouts