Removes Observer Blindening (#159)
## About The Pull Request

![filterbefore](https://github.com/f13babylon/f13babylon/assets/132588088/9a929c60-5788-4fe3-ac68-752b75ee52ab)

![after](https://github.com/f13babylon/f13babylon/assets/132588088/a1d34602-e290-4649-80d4-586ee2f2221c)

Removes the observer blindening system and allows ghosts/observers to
see items on the map. A proc that was restricted to admins as part of
this (as it was a way to cheese around it) was also restored.

## Why It's Good For The Game
This is a really interesting system inherited from Sunset, but is widely
disliked by many (including myself) and a suggestion to remove it
received unanimous praise.


![discord](https://github.com/f13babylon/f13babylon/assets/132588088/d39658bd-b5bb-4761-b93f-28c5406b6c7c)

I agree with this.

## Pre-Merge Checklist
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.

## Changelog
<!-- This is mostly optional for small Pull Requests, but if you value
being credited for your work in-game be sure to fill it out. To opt-out,
remove everything below including the start and end :cl: brackets. -->

<!-- If your Pull Request includes a minor single line variable edit,
probably do not fill out this changelog. -->
<!-- However, if your pull request includes massive or immediately
noticeable changes, briefly describe those changes in some way in this
changelog. -->

:cl:
del: Removes the observer blindening system.
/:cl: