Human + Anthro Ghouls and Super Mutants (#214)
## About The Pull Request

![p1_ghouls](https://github.com/f13babylon/f13babylon/assets/132588088/6e0f774f-6b64-4f99-95dc-e728ef13db47)

![p2_supermutants](https://github.com/f13babylon/f13babylon/assets/132588088/1dd439bc-70a4-4cd9-865c-699e082328c8)

![p3_behemoth](https://github.com/f13babylon/f13babylon/assets/132588088/0162e372-eab1-4a03-b736-8b8359758b8c)

Does something more interesting with #157 

Instead of removing the furry mob skins which are actually quite good, a
halfway solution so ₙₒₒₙₑ everyone is happy. There's now a 50% chance
whether a ghoul/supermutant is anthro or human, with the correct death
sprites to go along with it.

## Why It's Good For The Game
It makes sense to see both human and anthro ghouls/super mutants, and
some sprite variety for these common mobs is good. The alt skin variable
may be useful for other things in future too.

## Pre-Merge Checklist
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.

## Changelog
:cl:
tweak: Both human and anthro ghouls and super mutants will now appear.
/:cl: