Great Oak Mountain Retreat, the final BoS rebase to end them all. (#179)
## About The Pull Request

The complaints about "multi z" level combat and the tightness of the old
Bunker are heard. This PR adds Old Oak Mountain Retreat Facility, used
by the U.S Govt to establish control in the region in the event of
nuclear war. This facility has been built out of a mountain with a
bullet train that provides instantaneous transportation throughout the
sewer system.

This PR will be in parts, the lower level has a reactor, areas needs to
be set and other stuff too.


![image](https://github.com/f13babylon/f13babylon/assets/93557430/393f4ca8-f7e3-4c2c-8db5-19c33a495d95)


![image](https://github.com/f13babylon/f13babylon/assets/93557430/83473238-219e-4417-ba67-2d6aecdd141a)


![image](https://github.com/f13babylon/f13babylon/assets/93557430/54ee59fd-57cf-4364-8ae0-675a7c94164f)
(The above might be touched up at a later date, being rushed here!)

## Why It's Good For The Game

In between the aforementioned "multi z level" issue, and how the old
Bunker seriously needed a facelift, I thought I'll make this as an
interim for the new map. The innate lootspawns in this bunker have been
dumbed down but added utility has been added in place (plant disk/manip,
holsters, ec)

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [X] You tested this on a local server.
- [X] This code did not runtime during testing.
- [X] You documented all of your changes.

## Changelog

:cl:
Add: The final surface bunker to end all surface bunkers.
/:cl: