Enclave Marineification (#251)
## About The Pull Request

Changes all the icons and items of Enclave enlisted rank tabs (officer
tabs are the same between the Army and Marines) to be that of the Marine
Corps, while also adding some new ones that can be used for RP or
whatever. Also appropriately hands out said tabs. Private tab is GONE,
because it doesn't exist in the Marine Corps. You also may notice a
Petty Officer Third Class tab. What's that for? Well, it was originally
going to be for the Navy Corpsman loadout for the Specialist, however
loadouts seem to suffer with terminally low storage, so I wasn't able to
fit a rank tab into either of the loadouts. The item and icon are being
left in, however, just in case someone smarter than me can fix it.

Anyway, here are the sprites:

![sprites](https://github.com/f13babylon/f13babylon/assets/85942538/8b239726-1efb-43d3-a15d-731453df4a30)

## Why It's Good For The Game

We're Marines now! To hell with those Army tabs!
It's more immersive, and gosh darn it, that's all we Enclave players
have left. Immersion.

## Pre-Merge Checklist
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.


## Changelog


:cl:
add: Adds some brand new Marine Corps tabs.
del: Old Army tabs that aren't needed anymore.
tweak: Changes some of the old tabs to new ones.
imageadd: New tabs!
imagedel: Old ones are gone.
/:cl: