Chruch PR, attempt 3 to fix merge coflict (#249)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

<!-- NOTE: This format is NOT REQUIRED for things like runtime fixes,
code fixes and optimizations. In those instances you should try to give
a description of what is being fixed or optimized but are not required
to fill out the complete changelog. -->
<!-- Adjusting things like armor or weapon values for balance, while
they may be 'fixes' in their own way, are not considered code fixes as
described above and require the use of the Pull Request format below.-->


## About The Pull Request

this PR works on redesigning, and moving the warren church to bighorn

![2023-11-24 13 33
01](https://github.com/f13babylon/f13babylon/assets/118483925/f9e8cb97-cdb1-46b7-950a-4c0364fad0a8)

![2023-11-24 13 33
20](https://github.com/f13babylon/f13babylon/assets/118483925/3ee346c1-c788-4c87-8e9b-31381030bc31)


the preacher already is considered a bighorn citizen so the church being
in warren used to make no sense

this fixes the issue and adds a few additional community requested
changes


this is a second attempt because the last one was closed because of a
item that crashed the server that i forgot to mention in changelogs


### **please if you want something changed then say it**
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

Preacher is supposed to be a citizen of bighorn therefore have a church
in bighorn, but they for some reason had a church in warren instead of
bighorn, i have heard some community members arguing about it and
decided to act on it as it only makes sense for the church to be
somewhere close to bighorn instead of in warren.

adds more items for the priest to use, adds an actual spawning landmark
for the priest to spawn in the church instead of a random matrix

adds several other misc items such as a angel food cake, a holy water
bottle, bread, wine, and several spare bibles

oh also gives the preacher closet a shotgun with no ammo, since the
preacher landmark seems to have a shotgun icon.

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->


## Changelog
<!-- This is mostly optional for small Pull Requests, but if you value
being credited for your work in-game be sure to fill it out. To opt-out,
remove everything below including the start and end :cl: brackets. -->

<!-- If your Pull Request includes a minor single line variable edit,
probably do not fill out this changelog. -->
<!-- However, if your pull request includes massive or immediately
noticeable changes, briefly describe those changes in some way in this
changelog. -->

:cl:
add: church in (front of) bighorn
add: bible, bread, wine and water crate
add: holy water flask
add: angel food cake
add: Priest shotgun (with no ammo)
add: Preacher starter landmark
add: water bottles
tweak: changes the bible on the table to the whiskey bible
fix: fixes the prayer breads not working

/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->