Slightly changes bighorn bar's kitchen (#167)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

<!-- NOTE: This format is NOT REQUIRED for things like runtime fixes,
code fixes and optimizations. In those instances you should try to give
a description of what is being fixed or optimized but are not required
to fill out the complete changelog. -->
<!-- Adjusting things like armor or weapon values for balance, while
they may be 'fixes' in their own way, are not considered code fixes as
described above and require the use of the Pull Request format below.-->


## About The Pull Request
![2023-11-23 20 48
39](https://github.com/f13babylon/f13babylon/assets/118483925/b195c6bb-6559-4451-ad6f-fbdaef0e355c)
moves a door, adds kitchen vendor, changes the fridge subtypes so they
actually have required ingredients and meat in it
adds deepfrier
**adds a gibber, its behind the sign in map file so you cant see it**

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

several required changes to make the bighorn bar's kitchen way more
efficent than normal, such as moving the door to have direct acess to
the coldroom, changing the fridge types to add items, and a gibber

**once more the gibber is right there you cant see it because of the
damn sign**

![image](https://github.com/f13babylon/f13babylon/assets/118483925/a5d06420-d889-46f8-ade1-bfcef8e5bafd)



<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->


## Changelog
<!-- This is mostly optional for small Pull Requests, but if you value
being credited for your work in-game be sure to fill it out. To opt-out,
remove everything below including the start and end :cl: brackets. -->

<!-- If your Pull Request includes a minor single line variable edit,
probably do not fill out this changelog. -->
<!-- However, if your pull request includes massive or immediately
noticeable changes, briefly describe those changes in some way in this
changelog. -->

:cl:
add: bighorn deepfrier
add: bighorn kitchen tools vendor
add: bighorn bar backroom gibber
tweak: changes the subtypes of fridges in bighorn
tweak: moves the door in bighorn bar to be less annoying
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->