Surrender Indicator (#227)
finally, france pr

## About The Pull Request

Completely reworks how surrender functions.

## Why It's Good For The Game

Will still be available as an emote, however also now provides a status
effect, which is clickable, and allows for the user to surrender from a
fight at any given point in time. This also brings with it flick
overlay, which will allow for mutable appearances in emotes, as well as
changing the audio file for surrendering, adding sprites adjacent to the
surrender effect etc.

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->


## Changelog
<!-- This is mostly optional for small Pull Requests, but if you value
being credited for your work in-game be sure to fill it out. To opt-out,
remove everything below including the start and end :cl: brackets. -->

<!-- If your Pull Request includes a minor single line variable edit,
probably do not fill out this changelog. -->
<!-- However, if your pull request includes massive or immediately
noticeable changes, briefly describe those changes in some way in this
changelog. -->

:cl:
add: flick overlay
add: surrender status effect
imageadd: surrender sprites
soundadd: surrender sound effect
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Co-authored-by: Tom <8881105+tf-4@users.noreply.github.com>