NCR Gun Changes & Forgemaster Buff (#288)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

<!-- NOTE: This format is NOT REQUIRED for things like runtime fixes,
code fixes and optimizations. In those instances you should try to give
a description of what is being fixed or optimized but are not required
to fill out the complete changelog. -->
<!-- Adjusting things like armor or weapon values for balance, while
they may be 'fixes' in their own way, are not considered code fixes as
described above and require the use of the Pull Request format below.-->


## About The Pull Request

Re-balanced some weapons in NCR posession and gave Forgemaster the
ability to really be the Legion's counter-role to LO.

- Traded out Corporals R-82 for service carbine. Why? The 10mm is good
enough and the R-82 is an automatic gun and strong. Service carbine is a
fine alternative to a service rifle and actually gives more of a 'equal
choice'. (We have far too many round-start rifle caliber automatics.)
- Made the Republic's Pride not full-auto with stupidly high fire rates.
Still stornger than a normal M1 Garand
- Re-added SKS as a craftable gun for Forgemaster, rebalanced it to be a
slightly slower firing M1 but faster projectile speed than normal M1.
- Made certain crafting recipes, such as salvaged PA skin or service
rifles, rarer by making the LO the designated role to make those objects
easily without any blueprints needed.
- Fixed Forgemaster not getting a headset.

## Why It's Good For The Game

NCR's round start stuff was really good, and Legion needed some ways to
get similar stuff to NCR given the Forgemaster's role.

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [ ] You tested this on a local server.
- [ ] This code did not runtime during testing.
- [ ] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->


## Changelog
<!-- This is mostly optional for small Pull Requests, but if you value
being credited for your work in-game be sure to fill it out. To opt-out,
remove everything below including the start and end :cl: brackets. -->

<!-- If your Pull Request includes a minor single line variable edit,
probably do not fill out this changelog. -->
<!-- However, if your pull request includes massive or immediately
noticeable changes, briefly describe those changes in some way in this
changelog. -->

:cl:
add/remove: See the about section
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->